net/smc: Refactor buffer management for virtio-ism
This adds newly introduced dmb operations to main flow of SMC buffer
management.

The life cycles of local-allocated buffers (receive buffer, aka RMB)
are managed by local side, so it is necessary to allocate and free them
by local side.

And the send buffers are the peer-allocated RMBs actually, so make sure
attach or detach them on demand.

Signed-off-by: Tony Lu <tonylu@linux.alibaba.com>