net/smc: Copy message if shmem available
This introduces an optional path for SMC-D if shmem mode is available.
With this patch, it only needs once copy rather than twice copy in
original implementation.

In this design for shmem mode, sndbuf of SMC is actually the peer RMB
and it copies data to peer RMB from local user-space. This patch is
still an early approach, some uncritical pathes are not implemented.

Signed-off-by: Tony Lu <tonylu@linux.alibaba.com>